apps/backend/requirements.txt
Django>=5.2
djangorestframework>=3.15
drf-spectacular>=0.27
celery>=5.3
redis>=5.0
django-environ>=0.11
psycopg[binary]>=3.2
gunicorn>=21.2
